test(gates): apply code-review findings on the DOM harness (#5634)
Seven findings from a review of PR #5660, all in the new test/CI code —
no production source is touched.

P1 — `vi.waitFor` relied on its ~1s default (measured) while awaiting a
lazy `import('@/utils/export')`, whose first Vite transform on a cold
runner can approach that. `dom-tests` is in deploy-gate's required list,
so an intermittent timeout there would stall every PR in the repo, not
just this one. Both calls now pass an explicit 5s timeout.

P2 — `export-gate-wiring.test.mts` ran `resolveGateAction` for real with
`openBillingPortal` unmocked. Safe only because the cases exercised stop
short of the PAYMENT_ON_HOLD / RENEWAL_FAILED branches; adding either —
the obvious next step — would have fired a live portal-session fetch from
a unit test. Now mocked, with the constraint written down.

P2 — the runner glob matched `.test.mts` only, while `tests/` is 573
`.test.mjs` to 416 `.test.mts`. A contributor following the more common
convention would have added a file that never ran and never reported:
the exact silent-skip class this directory exists to prevent. Now
`*.test.{mts,mjs}`, verified by collecting a scratch `.mjs`.

P3 — dropped the dead `unstubGlobals` flag; removed `vi.restoreAllMocks()`
from two files so `restoreMocks: true` is the single cleanup mechanism;
split the harness into `helpers/i18n.mts` + `helpers/print-recorder.mts`
(they had no shared concern); and `clickExportPdf` now fails with a
message naming `printDelayMs` rather than an opaque empty-array mismatch
when the mirrored timing constants go stale.

Re-verified after the changes: 86/86 pass, typecheck clean (it caught an
`afterEach` import the removal orphaned — biome did not), lint clean, and
the full 79-mutation sweep still catches 79/79 with zero survivors.
